Jingwu/SPICE-Detector/py_detector/requirements.txt
numpy
scipy
argparse
pyyaml